Move solveResult to seperate file
